update getting cache URL
Update getting the cache URL in the
`Cache_Enabler_Disk::get_cache_url()` method. We no longer need to
allow a default value if the parameter given is empty due to the
changes introduced in PR #251 (`user_trailingslashit( home_url() )` can
be called instead for the same result).

This updates what was originally added in PR #237.