update conditional statements (#264)
Update conditional statements checking for a numeric user ID to be an
`elseif` instead of a separate `if` statement. There is no reason to
check this if the value was empty beforehand and set as the current
user `WP_User` instance. This updates what was added in PR #247.

Update conditional statement in the directory object filter to be an
`elseif` statement instead of an `else` statement. Using a ternary
operator makes no sense here. This updates what was added in PR #252.