fix WebP URL conversion edge case (#275)
Fix a rare edge case in the WebP URL conversion. This method needs to
return the full match if an image is not found. This has not been an
issue due to the image URL regex always finding an image. However, if
for some reason an image was not found nothing would have been
returned, resulting in whatever was matched not being sent back. That
would result in the static HTML file created missing what was matched,
breaking the layout.