update advanced-cache.php handling (#283)
Rename the `advanced-cache-sample.php` file to `advanced-cache.php` for
backward compatibility when updating from a couple old versions with a
certain setting enabled. More specifically, any version from 1.2.3 (may
be earlier) to 1.3.5 and the "Clear the complete cache if any plugin
has been upgraded" setting was enabled. If this is not done then a PHP
warning will occur due to a file not existing when trying to be copied.
I believe this will be better than the alternative of leaving an empty
`advanced-cache.php` file and then leaving
`advanced-cache-sample.php`.

This updates the changes made in PR #260.