update index file handling
Update the `Cache_Enabler_Engine::is_index()` method to use the
`$_SERVER['SCRIPT_NAME']` again like it was before PR #260 as this is
in the CGI/1.1 specification, making it a more reliable superglobal.
This method will now check if the `CACHE_ENABLER_INDEX_FILE` constant
ends with `$_SERVER['SCRIPT_NAME']` instead of if it equals
`$_SERVER['SCRIPT_FILENAME']`.