update advanced-cache.php handling (#297)
Update the `advanced-cache.php` file creation handling to use the new
`CACHE_ENABLER_CONSTANTS_FILE` constant discussed in #296 (thanks
@nlemoine).

Update the `advanced-cache.php` file to delete itself if considered
stale or abandoned. It would be stale if the
`CACHE_ENABLER_CONSTANTS_FILE` has changed after the
`wp-content/advanced-cache.php` file was created (like in an example
shared #296). It would be abandoned if the plugin was manually deleted,
like through an FTP client. If stale, the file will be created again
with the new value, which currently takes place during the requirements
check.

Update the `CACHE_ENABLER_MIN_WP` constant to `5.1`, which should have
been updated in PR #295.