inc/cache_enabler_engine.class.php: speed up is_cacheable()
The is_cacheable() function is checking,

  if ( $has_html_tag && $has_html_doctype && ! $has_xsl_stylesheet )

before returning true. PHP does short-circuit its conditionals, but
the expensive part of that statement is computing the values of the
variables themselves, and they are all computed before the "if"
statement is run.

To speed up the check a bit, that conditional has been refacored into
nested "if" statements. The variable $has_html_doctype, for example,
is now computed only if $has_html_tag is true. This should be a small
improvement for non-cacheable resources, since the function will
return false sooner.