Support more (html4, xhtml1.0, xhtml1.1) doctypes (#338)
* inc/cache_enabler_engine.class.php: support more DOCTYPEs

The is_cacheable() function currently checks for an html5 doctype when
deciding whether or not a given resource is cacheable. Some sites
however may still be using xhtml1.1, xhtml1.0, or even html4. This
commit updates the check to support those additional doctypes.

Closes: https://github.com/keycdn/cache-enabler/issues/324

* inc/cache_enabler_engine.class.php: speed up is_cacheable()

The is_cacheable() function is checking,

  if ( $has_html_tag && $has_html_doctype && ! $has_xsl_stylesheet )

before returning true. PHP does short-circuit its conditionals, but
the expensive part of that statement is computing the values of the
variables themselves, and they are all computed before the "if"
statement is run.

To speed up the check a bit, that conditional has been refacored into
nested "if" statements. The variable $has_html_doctype, for example,
is now computed only if $has_html_tag is true. This should be a small
improvement for non-cacheable resources, since the function will
return false sooner.

* inc/cache_enabler_engine.class.php: fix PHP parser confusion

Apparently, PHP really hates to see <? ... ?> tags, even inside of
comments. This commit rewrites a comment to avoid those tags, instead
describing them in words. It fixes the not-obviously-related error,

  PHP Parse error:  Unclosed '{' on line 272 in /usr/local/lib/wordpress/
  plugins/cache-enabler/inc/cache_enabler_engine.class.php on line 483